LPS-109473 Merging both wrappers
