+++ modules/apps/depot/depot-web/src/main/java/com/liferay/depot/web/internal/asset/model/DepotAssetRendererFactoryWrapper.java
@@ -18,10 +18,19 @@ import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetRenderer;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.asset.kernel.model.ClassTypeReader;
+import com.liferay.depot.service.DepotEntryLocalService;
+import com.liferay.depot.web.internal.application.controller.DepotApplicationController;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.GroupConstants;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.security.permission.PermissionChecker;
+import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.service.ServiceContext;
+import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.GroupThreadLocal;
 
 import java.util.Locale;
 
@@ -31,8 +40,20 @@ import javax.portlet.WindowState;
 /**
  * @author Adolfo Pérez
  */
-public abstract class DepotAssetRendererFactoryWrapper<T>
+public class DepotAssetRendererFactoryWrapper<T>
 	implements AssetRendererFactory<T> {
+
+	public DepotAssetRendererFactoryWrapper(
+		AssetRendererFactory assetRendererFactory,
+		DepotApplicationController depotApplicationController,
+		DepotEntryLocalService depotEntryLocalService,
+		GroupLocalService groupLocalService) {
+
+		_assetRendererFactory = assetRendererFactory;
+		_depotApplicationController = depotApplicationController;
+		_depotEntryLocalService = depotEntryLocalService;
+		_groupLocalService = groupLocalService;
+	}
 
 	@Override
 	public AssetEntry getAssetEntry(long assetEntryId) throws PortalException {
@@ -86,6 +107,12 @@ public abstract class DepotAssetRendererFactoryWrapper<T>
 
 	@Override
 	public ClassTypeReader getClassTypeReader() {
+		if (isSelectable()) {
+			return new DepotClassTypeReader(
+				_assetRendererFactory.getClassTypeReader(),
+				_depotEntryLocalService);
+		}
+
 		return getAssetRendererFactory().getClassTypeReader();
 	}
 
@@ -180,6 +207,15 @@ public abstract class DepotAssetRendererFactoryWrapper<T>
 
 	@Override
 	public boolean isSelectable() {
+		Group group = _getGroup();
+
+		if ((group != null) && (group.getType() == GroupConstants.TYPE_DEPOT) &&
+			!_depotApplicationController.isClassNameEnabled(
+				getClassName(), group.getGroupId())) {
+
+			return false;
+		}
+
 		return getAssetRendererFactory().isSelectable();
 	}
 
@@ -198,6 +234,30 @@ public abstract class DepotAssetRendererFactoryWrapper<T>
 		getAssetRendererFactory().setPortletId(portletId);
 	}
 
-	protected abstract AssetRendererFactory<T> getAssetRendererFactory();
+	protected AssetRendererFactory getAssetRendererFactory() {
+		return _assetRendererFactory;
+	}
+
+	private Group _getGroup() {
+		ServiceContext serviceContext =
+			ServiceContextThreadLocal.getServiceContext();
+
+		if (serviceContext == null) {
+			return _groupLocalService.fetchGroup(GroupThreadLocal.getGroupId());
+		}
+
+		ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
+
+		if (themeDisplay != null) {
+			return themeDisplay.getScopeGroup();
+		}
+
+		return _groupLocalService.fetchGroup(serviceContext.getScopeGroupId());
+	}
+
+	private final AssetRendererFactory _assetRendererFactory;
+	private final DepotApplicationController _depotApplicationController;
+	private final DepotEntryLocalService _depotEntryLocalService;
+	private final GroupLocalService _groupLocalService;
 
 }
